+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 -----
 - Add bioservices python package to the ``resolwebio/common:2.8.0``
   Docker image
+- Add ``upload-idat`` process
 
 
 ===================